Consolidate modules into model
