--- config.py
@@ -1,28 +1,0 @@
-from model import LanguageModelConfig, TransformerConfig
-
-
-def default_config() -> LanguageModelConfig:
-    """Return the default SLNCX model configuration."""
-    return LanguageModelConfig(
-        vocab_size=128 * 1024,
-        pad_token=0,
-        eos_token=2,
-        sequence_len=8192,
-        embedding_init_scale=1.0,
-        output_multiplier_scale=0.5773502691896257,
-        embedding_multiplier_scale=78.38367176906169,
-        model=TransformerConfig(
-            emb_size=48 * 128,
-            widening_factor=8,
-            key_size=128,
-            num_q_heads=48,
-            num_kv_heads=8,
-            num_layers=64,
-            attn_output_multiplier=0.08838834764831845,
-            shard_activations=True,
-            num_experts=8,
-            num_selected_experts=2,
-            data_axis="data",
-            model_axis="model",
-        ),
-    )
